Python Notebooks
The first three python notebooks are "tutorial"-based and give an introduction to python commands and its general use as a programming language; the pandas notebooks provide a guide/introduction to explaining through what the pandas package is within python and how to use it to manipulate data. The 4th notebook titled "TeenCountyDataAlaska" contains the data that this repository is centered around.